+++ utils/handlers.py
@@ -1,0 +1,29 @@
+from telegram import (
+    ReplyKeyboardMarkup,
+)
+
+from telegram.ext import (
+    ConversationHandler,
+)
+
+
+def maxNumberReached(update, context):
+    update.message.reply_text(
+        "Hey! Thanks for your interest, but it seems that the maximum number of users has been reached."
+    )
+    return ConversationHandler.END
+
+
+def botStopped(update, context):
+    update.message.reply_text(
+        "The airdrop has been completed. Thanks for you interest."
+    )
+    return ConversationHandler.END
+
+
+def botPaused(update, context):
+    update.message.reply_text(
+        "The airdrop has been temporarily paused, please try again later",
+        reply_markup=ReplyKeyboardMarkup([["/start"]]),
+    )
+    return ConversationHandler.END
